Added voice calls and fixed a few bugs

+++ lib/ui/views/parent/home_view.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:tunza_app/core/enums/viewstate.dart';
+import 'package:tunza_app/core/models/child.dart';
 import 'package:tunza_app/core/viewmodels/parent/home_model.dart';
 import 'package:tunza_app/ui/views/base_view.dart';
 import 'package:tunza_app/ui/views/parent/communicationview.dart';
@@ -39,7 +40,7 @@ class _HomeViewState extends State<HomeView>{
 //              }),
 
               IconButton(icon: Icon(Icons.search), onPressed: (){
-                showSearch(context: context, delegate: AppSearch());
+                showSearch(context: context, delegate: AppSearch(model.childList));
               })
               
 
@@ -129,9 +130,55 @@ class _HomeViewState extends State<HomeView>{
                   ),
                   title: Text(model.childList[i].child_name),
                   subtitle: Text(model.childList[i].child_date_of_birth),
-                  trailing: Icon(Icons.remove_red_eye, size: 10,),
+                  trailing: model.state==ViewState.Edit?Column(
+                  mainAxisSize: MainAxisSize.max,
+                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  children: <Widget>[
+                    GestureDetector(
+
+                      child: Icon(Icons.edit,size: 16,color: Colors.green,),
+                      onTap: (){
+                        Navigator.pushNamed(context, "add_child",arguments: [model.childList[i]]);
+                      },
+                    ),
+                    GestureDetector(
+                      child: Icon(Icons.delete_forever,size: 16,color: Colors.red,),
+                      onTap: (){
+                        model.setState(ViewState.Delete);
+                        showDialog(context: context,
+                            builder: (context){
+                              return AlertDialog(
+                                title: Text("Delete Confirmation"),
+                                content: Text("You are about to delete your child's information. This action is not reversible."),
+                                actions: <Widget>[
+
+                                  FlatButton(child: Text("Cancel"),
+                                    onPressed: (){
+                                      model.setState(ViewState.Edit);
+                                      Navigator.pop(context);
+                                    },
+                                    color: Colors.blueGrey[50],
+                                  ),
+                                  FlatButton(
+                                    child: Text("Delete"),
+                                    onPressed: ()async{
+                                      Navigator.pop(context);
+                                      await model.deleteChild(model.childList[i].child_id);
+                                    },color: Colors.red,
+                                  ),
+
+                                ],
+
+                              );
+                            });
+                      },
+
+                    )
+                  ],
+                ):Icon(Icons.remove_red_eye,size: 16),
                   onLongPress: ()async{
-                    await model.deleteChild(model.childList[i].child_id);
+                    //@todo: edit delete stuff here
+                    model.state!=ViewState.Edit?model.setState(ViewState.Edit):model.setState(ViewState.Idle);
                   },
                   onTap: (){
                     Navigator.pushNamed(context, "child",arguments:model.childList[i]);
@@ -169,7 +216,7 @@ class _HomeViewState extends State<HomeView>{
             label: Text("add"),
             tooltip: "add child",
             onPressed: (){
-              Navigator.pushNamed(context, "add_child");
+              Navigator.pushNamed(context, "add_child",arguments: []);
             },
           ):null,
 
@@ -180,19 +227,9 @@ class _HomeViewState extends State<HomeView>{
 
 }
 class AppSearch extends SearchDelegate<String>{
-
-final cities = [
-"Otse",
-"Kanye",
-"Lobatse",
-"Ramotswa",
-"Zilapo"
-];
- final recentCities =[
-  "Otse",
- "Kanye" 
- ];
-
+  List<Child> childList=[];
+  final recentSearch =[];
+AppSearch(this.childList);
   @override
   List<Widget> buildActions(BuildContext context) {
     //what kind of actions do i want to perform
@@ -239,28 +276,28 @@ final cities = [
   @override
   Widget buildSuggestions(BuildContext context) {
     // give suggestions whan someone searches something
-    final suggestionList = query.isEmpty ? recentCities : cities.where((p)=>p.startsWith(query)).toList();
-    return ListView.builder(
-      itemBuilder:(context, index) => ListTile(
-        onTap: (){
-          showResults(context);
-        },
-        leading: Icon(Icons.location_city),
-        title: RichText(
-          text: TextSpan(
-            text: suggestionList[index].substring(0, query.length),
-            style: TextStyle(
-              color: Colors.black,fontWeight: FontWeight.bold
-            ),
-            children: [TextSpan(
-              text: suggestionList[index].substring(query.length),
-              style: TextStyle(color: Colors.grey)
-            )]
-          )
+    final suggestionList = query.isEmpty ? recentSearch : this.childList.where((p)=>p.child_name.startsWith(query)).toList();
+    return ListView.builder(itemCount:suggestionList.length ,itemBuilder: (context,i){
+
+      return Card(
+        child: ListTile(
+          leading: CircleAvatar(
+            backgroundImage: NetworkImage("https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcQOL0wCKxBI3i8-S8OnckAeaszWpVHziXKSnzBRNuABcKZw66M-&usqp=CAU"),
+          ),
+          title: Text(suggestionList[i].child_name),
+          subtitle: Text(suggestionList[i].child_date_of_birth),
+          trailing: Icon(Icons.remove_red_eye,size: 16),
+          onLongPress: ()async{
+            //@todo: edit delete stuff here
+
+          },
+          onTap: (){
+
+            Navigator.pushReplacementNamed(context, "child",arguments:suggestionList[i]);
+          },
         ),
-        ),
-        itemCount: suggestionList.length,
-    );
+      );
+    });
   }
 
 }
